+++ scripts/sync.sh
@@ -12,7 +12,6 @@ cd "$(dirname "$0")" || { echo "Could not cd into base directory"; exit 1; }
 base="$(pwd)"
 
 rust="$1"
-alloc="library/alloc"
 
 current_file="${base}/sync-current"
 current_rev="$(<"${current_file}")" || { echo "Invalid current revision"; exit 1; }
@@ -25,13 +24,13 @@ current="$(git rev-parse "${current_rev}")" \
 new="$(git rev-parse "${new_rev}")" \
     || { echo "New git revision (${new_rev}) not found"; exit 1; }
 
-git diff "${current}" "${new}" "${alloc}/src/collections/binary_heap/mod.rs" \
+git diff "${current}" "${new}" "library/alloc/src/collections/binary_heap/mod.rs" \
     | patch --merge=diff3 "${base}/../src/lib.rs"
-git diff "${current}" "${new}" "${alloc}/src/testing/crash_test.rs" \
+git diff "${current}" "${new}" "library/alloctests/testing/crash_test.rs" \
     | patch --merge=diff3 "${base}/../tests/binary_heap/crash_test.rs"
-git diff "${current}" "${new}" "${alloc}/src/collections/binary_heap/tests.rs" \
+git diff "${current}" "${new}" "library/alloctests/tests/collections/binary_heap.rs" \
     | patch --merge=diff3 "${base}/../tests/binary_heap.rs"
-git diff "${current}" "${new}" "${alloc}/benches/binary_heap.rs" \
+git diff "${current}" "${new}" "library/alloctests/benches/binary_heap.rs" \
     | patch --merge=diff3 "${base}/../benches/upstream/binary_heap.rs"
 
 echo "${new}" > "${current_file}"